backend/requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
python-decouple==3.6
Pillow==9.0.1